refactor: 값 객체는 value 패키지로 이동
