--- src/main/java/com/birca/bircabackend/command/birca/domain/Visibility.java
@@ -1,5 +1,0 @@
-package com.birca.bircabackend.command.birca.domain;
-
-public enum Visibility {
-    PUBLIC, PRIVATE
-}
