+++ src/main/java/com/birca/bircabackend/command/birca/domain/value/SpecialGoodsStockState.java
@@ -1,4 +1,4 @@
-package com.birca.bircabackend.command.birca.domain;
+package com.birca.bircabackend.command.birca.domain.value;
 
 public enum SpecialGoodsStockState {
