+++ src/main/java/com/birca/bircabackend/command/birca/domain/value/Visibility.java
@@ -1,0 +1,5 @@
+package com.birca.bircabackend.command.birca.domain.value;
+
+public enum Visibility {
+    PUBLIC, PRIVATE
+}
