+++ src/main/java/com/birca/bircabackend/command/birca/dto/BirthdayCafeCreateRequest.java
@@ -1,8 +1,8 @@
 package com.birca.bircabackend.command.birca.dto;
 
 import com.birca.bircabackend.command.birca.domain.BirthdayCafe;
-import com.birca.bircabackend.command.birca.domain.Schedule;
-import com.birca.bircabackend.command.birca.domain.Visitants;
+import com.birca.bircabackend.command.birca.domain.value.Schedule;
+import com.birca.bircabackend.command.birca.domain.value.Visitants;
 
 import java.time.LocalDateTime;
