+++ src/test/java/com/birca/bircabackend/command/birca/application/BirthdayCafeServiceTest.java
@@ -2,6 +2,7 @@ package com.birca.bircabackend.command.birca.application;
 
 import com.birca.bircabackend.command.auth.authorization.LoginMember;
 import com.birca.bircabackend.command.birca.domain.*;
+import com.birca.bircabackend.command.birca.domain.value.*;
 import com.birca.bircabackend.command.birca.dto.BirthdayCafeCreateRequest;
 import com.birca.bircabackend.command.birca.exception.BirthdayCafeErrorCode;
 import com.birca.bircabackend.common.exception.BusinessException;
